Audit logs QOL improvements
