+++ workos/resources/audit_logs.py
@@ -15,38 +15,3 @@ class AuditLogExport(WorkOSModel):
     updated_at: str
     state: AuditLogExportState
     url: Optional[str] = None
-
-
-class AuditLogEventActor(TypedDict):
-    """Describes the entity that generated the event."""
-
-    id: str
-    metadata: NotRequired[dict]
-    name: NotRequired[str]
-    type: str
-
-
-class AuditLogEventTarget(TypedDict):
-    """Describes the entity that was targeted by the event."""
-
-    id: str
-    metadata: NotRequired[dict]
-    name: NotRequired[str]
-    type: str
-
-
-class AuditLogEventContext(TypedDict):
-    """Attributes of audit log event context."""
-
-    location: str
-    user_agent: NotRequired[str]
-
-
-class AuditLogEvent(TypedDict):
-    action: str
-    version: NotRequired[int]
-    occurred_at: str  # ISO-8601 datetime of when an event occurred
-    actor: AuditLogEventActor
-    targets: Sequence[AuditLogEventTarget]
-    context: AuditLogEventContext
-    metadata: NotRequired[dict]
